+++ controllers/doguInstallManager_test.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	"errors"
+	"github.com/cloudogu/k8s-dogu-operator/controllers/util"
 	"testing"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -12,8 +13,6 @@ import (
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd/api"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
@@ -105,54 +104,26 @@ func getDoguInstallManagerTestData(t *testing.T) (*k8sv1.Dogu, *core.Dogu, *core
 
 func TestNewDoguInstallManager(t *testing.T) {
 	// override default controller method to retrieve a kube config
-	oldGetConfigOrDieDelegate := ctrl.GetConfigOrDie
-	defer func() { ctrl.GetConfigOrDie = oldGetConfigOrDieDelegate }()
-	ctrl.GetConfigOrDie = func() *rest.Config {
-		return &rest.Config{}
-	}
+	oldGetConfigDelegate := ctrl.GetConfig
+	defer func() { ctrl.GetConfig = oldGetConfigDelegate }()
+	ctrl.GetConfig = createTestRestConfig
 
 	t.Run("success", func(t *testing.T) {
 		// given
-		myClient := fake.NewClientBuilder().WithScheme(runtime.NewScheme()).Build()
+		myClient := fake.NewClientBuilder().WithScheme(getTestScheme()).Build()
 		operatorConfig := &config.OperatorConfig{}
 		operatorConfig.Namespace = "test"
-		cesRegistry := cesmocks.NewRegistry(t)
 		doguRegistry := cesmocks.NewDoguRegistry(t)
-		globalConfig := cesmocks.NewConfigurationContext(t)
-		eventRecorder := extMocks.NewEventRecorder(t)
+		cesRegistry := cesmocks.NewRegistry(t)
 		cesRegistry.On("DoguRegistry").Return(doguRegistry)
-		cesRegistry.On("GlobalConfig").Return(globalConfig)
-
-		// when
-		doguManager, err := NewDoguInstallManager(myClient, operatorConfig, cesRegistry, eventRecorder)
-
-		// then
-		require.NoError(t, err)
-		require.NotNil(t, doguManager)
-	})
-
-	t.Run("fail when creating client", func(t *testing.T) {
-		// given
-
-		// override default controller method to return a config that fail the client creation
-		oldGetConfigOrDieDelegate := ctrl.GetConfigOrDie
-		defer func() { ctrl.GetConfigOrDie = oldGetConfigOrDieDelegate }()
-		ctrl.GetConfigOrDie = func() *rest.Config {
-			return &rest.Config{ExecProvider: &api.ExecConfig{}, AuthProvider: &api.AuthProviderConfig{}}
-		}
-
-		myClient := fake.NewClientBuilder().WithScheme(runtime.NewScheme()).Build()
-		operatorConfig := &config.OperatorConfig{}
-		operatorConfig.Namespace = "test"
-		cesRegistry := &cesmocks.Registry{}
+		mgrSet := &util.ManagerSet{}
 		eventRecorder := extMocks.NewEventRecorder(t)
 
 		// when
-		doguManager, err := NewDoguInstallManager(myClient, operatorConfig, cesRegistry, eventRecorder)
+		doguManager := NewDoguInstallManager(myClient, operatorConfig, cesRegistry, mgrSet, eventRecorder)
 
 		// then
-		require.Error(t, err)
-		require.Nil(t, doguManager)
+		require.NotNil(t, doguManager)
 	})
 }