+++ controllers/upgrade/upgradeExecutor_test.go
@@ -55,16 +55,21 @@ func TestNewUpgradeExecutor(t *testing.T) {
 		saCreator := mocks.NewServiceAccountCreator(t)
 		k8sFileEx := mocks.NewFileExtractor(t)
 		applier := mocks.NewCollectApplier(t)
-		doguRegistry := regmock.NewDoguRegistry(t)
-		globalConfig := regmock.NewConfigurationContext(t)
-		mockRegistry := regmock.NewRegistry(t)
-		mockRegistry.On("DoguRegistry").Return(doguRegistry, nil)
-		mockRegistry.On("GlobalConfig").Return(globalConfig)
 		eventRecorder := extMocks.NewEventRecorder(t)
 		commandExecutor := mocks.NewCommandExecutor(t)
+		mockUpserter := mocks.NewResourceUpserter(t)
+		mgrSet := &util.ManagerSet{
+			RestConfig:            testRestConfig,
+			ImageRegistry:         imageRegMock,
+			ServiceAccountCreator: saCreator,
+			FileExtractor:         k8sFileEx,
+			CollectApplier:        applier,
+			CommandExecutor:       commandExecutor,
+			ResourceUpserter:      mockUpserter,
+		}
 
 		// when
-		actual := NewUpgradeExecutor(myClient, testRestConfig, commandExecutor, eventRecorder, imageRegMock, applier, k8sFileEx, saCreator, mockRegistry)
+		actual := NewUpgradeExecutor(myClient, mgrSet, eventRecorder)
 
 		// then
 		require.NotNil(t, actual)