+++ internal/cloudogu/mocks/DoguResourceGenerator.go
@@ -138,6 +138,61 @@ func (_c *DoguResourceGenerator_CreateDoguPVC_Call) RunAndReturn(run func(*v1.Do
 	return _c
 }
 
+// CreateDoguSecret provides a mock function with given fields: doguResource, stringData
+func (_m *DoguResourceGenerator) CreateDoguSecret(doguResource *v1.Dogu, stringData map[string]string) (*corev1.Secret, error) {
+	ret := _m.Called(doguResource, stringData)
+
+	var r0 *corev1.Secret
+	var r1 error
+	if rf, ok := ret.Get(0).(func(*v1.Dogu, map[string]string) (*corev1.Secret, error)); ok {
+		return rf(doguResource, stringData)
+	}
+	if rf, ok := ret.Get(0).(func(*v1.Dogu, map[string]string) *corev1.Secret); ok {
+		r0 = rf(doguResource, stringData)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*corev1.Secret)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(*v1.Dogu, map[string]string) error); ok {
+		r1 = rf(doguResource, stringData)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
+// DoguResourceGenerator_CreateDoguSecret_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'CreateDoguSecret'
+type DoguResourceGenerator_CreateDoguSecret_Call struct {
+	*mock.Call
+}
+
+// CreateDoguSecret is a helper method to define mock.On call
+//   - doguResource *v1.Dogu
+//   - stringData map[string]string
+func (_e *DoguResourceGenerator_Expecter) CreateDoguSecret(doguResource interface{}, stringData interface{}) *DoguResourceGenerator_CreateDoguSecret_Call {
+	return &DoguResourceGenerator_CreateDoguSecret_Call{Call: _e.mock.On("CreateDoguSecret", doguResource, stringData)}
+}
+
+func (_c *DoguResourceGenerator_CreateDoguSecret_Call) Run(run func(doguResource *v1.Dogu, stringData map[string]string)) *DoguResourceGenerator_CreateDoguSecret_Call {
+	_c.Call.Run(func(args mock.Arguments) {
+		run(args[0].(*v1.Dogu), args[1].(map[string]string))
+	})
+	return _c
+}
+
+func (_c *DoguResourceGenerator_CreateDoguSecret_Call) Return(_a0 *corev1.Secret, _a1 error) *DoguResourceGenerator_CreateDoguSecret_Call {
+	_c.Call.Return(_a0, _a1)
+	return _c
+}
+
+func (_c *DoguResourceGenerator_CreateDoguSecret_Call) RunAndReturn(run func(*v1.Dogu, map[string]string) (*corev1.Secret, error)) *DoguResourceGenerator_CreateDoguSecret_Call {
+	_c.Call.Return(run)
+	return _c
+}
+
 // CreateDoguService provides a mock function with given fields: doguResource, imageConfig
 func (_m *DoguResourceGenerator) CreateDoguService(doguResource *v1.Dogu, imageConfig *pkgv1.ConfigFile) (*corev1.Service, error) {
 	ret := _m.Called(doguResource, imageConfig)
@@ -189,6 +244,61 @@ func (_c *DoguResourceGenerator_CreateDoguService_Call) Return(_a0 *corev1.Servi
 }
 
 func (_c *DoguResourceGenerator_CreateDoguService_Call) RunAndReturn(run func(*v1.Dogu, *pkgv1.ConfigFile) (*corev1.Service, error)) *DoguResourceGenerator_CreateDoguService_Call {
+	_c.Call.Return(run)
+	return _c
+}
+
+// GetPodTemplate provides a mock function with given fields: doguResource, dogu
+func (_m *DoguResourceGenerator) GetPodTemplate(doguResource *v1.Dogu, dogu *core.Dogu) (*corev1.PodTemplateSpec, error) {
+	ret := _m.Called(doguResource, dogu)
+
+	var r0 *corev1.PodTemplateSpec
+	var r1 error
+	if rf, ok := ret.Get(0).(func(*v1.Dogu, *core.Dogu) (*corev1.PodTemplateSpec, error)); ok {
+		return rf(doguResource, dogu)
+	}
+	if rf, ok := ret.Get(0).(func(*v1.Dogu, *core.Dogu) *corev1.PodTemplateSpec); ok {
+		r0 = rf(doguResource, dogu)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*corev1.PodTemplateSpec)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(*v1.Dogu, *core.Dogu) error); ok {
+		r1 = rf(doguResource, dogu)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
+// DoguResourceGenerator_GetPodTemplate_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'GetPodTemplate'
+type DoguResourceGenerator_GetPodTemplate_Call struct {
+	*mock.Call
+}
+
+// GetPodTemplate is a helper method to define mock.On call
+//   - doguResource *v1.Dogu
+//   - dogu *core.Dogu
+func (_e *DoguResourceGenerator_Expecter) GetPodTemplate(doguResource interface{}, dogu interface{}) *DoguResourceGenerator_GetPodTemplate_Call {
+	return &DoguResourceGenerator_GetPodTemplate_Call{Call: _e.mock.On("GetPodTemplate", doguResource, dogu)}
+}
+
+func (_c *DoguResourceGenerator_GetPodTemplate_Call) Run(run func(doguResource *v1.Dogu, dogu *core.Dogu)) *DoguResourceGenerator_GetPodTemplate_Call {
+	_c.Call.Run(func(args mock.Arguments) {
+		run(args[0].(*v1.Dogu), args[1].(*core.Dogu))
+	})
+	return _c
+}
+
+func (_c *DoguResourceGenerator_GetPodTemplate_Call) Return(_a0 *corev1.PodTemplateSpec, _a1 error) *DoguResourceGenerator_GetPodTemplate_Call {
+	_c.Call.Return(_a0, _a1)
+	return _c
+}
+
+func (_c *DoguResourceGenerator_GetPodTemplate_Call) RunAndReturn(run func(*v1.Dogu, *core.Dogu) (*corev1.PodTemplateSpec, error)) *DoguResourceGenerator_GetPodTemplate_Call {
 	_c.Call.Return(run)
 	return _c
 }